+++ packages/nextjs/app/poolAdreess/page.tsx
@@ -3,7 +3,7 @@
 import { useState } from "react";
 import Image from "next/image";
 import { CiCalendar } from "react-icons/ci";
-import BorrowContentModal from "~~/components/BorrowContentModal/BorrowContentModal";
+import BorrowRepayContentModal from "~~/components/BorrowRepayContentModal/BorrowRepayContentModal";
 import Button from "~~/components/Button/Button";
 import DepositContentModal from "~~/components/DepositContentModal/DepositContentModal";
 import Analytics from "~~/components/Graph/Analytics";
@@ -162,7 +162,7 @@ function Page() {
           </div>
         </div>
       </div>
-      {isModalOpen && <BorrowContentModal setIsModalOpen={() => setIsModalOpen(false)}></BorrowContentModal>}
+      {isModalOpen && <BorrowRepayContentModal setIsModalOpen={() => setIsModalOpen(false)}></BorrowRepayContentModal>}
     </section>
   );
 }
