+++ packages/nextjs/styles/globals.css
@@ -165,7 +165,7 @@ input{
   border-radius: 5px;
   background: #19192B;
   padding-block: 2rem;
-  padding-inline: 3rem;
+  padding-inline: 2rem;
   z-index: 1;
   color: #ffffff;
 }
